+++ src/Avalonia.X11/Vulkan/VulkanSupport.cs
@@ -69,18 +69,18 @@ internal class VulkanSupport
 
         public double Scaling => _handle.Scaling;
         public PixelSize Size => _handle.Size;
-        public IntPtr CreateSurface(IVulkanPlatformGraphicsContext context) =>
+        public ulong CreateSurface(IVulkanPlatformGraphicsContext context) =>
             CreateXlibSurface(_display, _handle.Handle, context.Instance);
     }
 
-    private static IntPtr CreateDummySurface(IntPtr display, IVulkanInstance instance)
+    private static ulong CreateDummySurface(IntPtr display, IVulkanInstance instance)
     {
         var surf = CreateXlibSurface(display, s_offscreenWindow, instance);
         XLib.XSync(display, true);
         return surf;
     }
 
-    private static IntPtr CreateXlibSurface(IntPtr display, IntPtr window, IVulkanInstance instance)
+    private static ulong CreateXlibSurface(IntPtr display, IntPtr window, IVulkanInstance instance)
     {
         var vulkanXlib = new X11VulkanInterface(instance);
         var createInfo = new VkXlibSurfaceCreateInfoKHR
